move unit tests into test/unit/
