+++ test/unit/invitation.js
@@ -1,23 +1,23 @@
 const expect = require('chai').expect;
-const Permission = require('../schema/permission');
+const Invitation = require('../../schema/invitation');
 const ValidationError = require('objection/lib/model/ValidationError');
 
-describe('Permission', () => {
+describe('Invitation', () => {
   it('throws a validation error when required properties are missing', () => {
     const badJson = {
-      establishmentId: 'abcd-1234',
+      establishmentId: 100,
       profileId: '1234567'
     };
-    expect(() => Permission.fromJson(badJson)).to.throw(ValidationError, /required/);
+    expect(() => Invitation.fromJson(badJson)).to.throw(ValidationError, /required/);
   });
 
   it('throws a validation error when invalid values are provided', () => {
     const badJson = {
       role: 'super',
-      establishmentId: 'abcd-1234',
+      establishmentId: 100,
       profileId: '1234567'
     };
-    expect(() => Permission.fromJson(badJson)).to.throw(ValidationError, /allowed values/);
+    expect(() => Invitation.fromJson(badJson)).to.throw(ValidationError, /allowed values/);
   });
 
   it('throws a validation error when unknown properties are provided', () => {
@@ -27,7 +27,7 @@ describe('Permission', () => {
       profileId: '1234567',
       unknown: 'example'
     };
-    expect(() => Permission.fromJson(badJson)).to.throw(ValidationError, /invalid additional property/);
+    expect(() => Invitation.fromJson(badJson)).to.throw(ValidationError, /invalid additional property/);
   });
 
   it('successfully instantiates when given a valid schema', () => {
@@ -36,6 +36,6 @@ describe('Permission', () => {
       establishmentId: 100,
       profileId: '1234567'
     };
-    expect(Permission.fromJson(goodJson)).to.be.an('object');
+    expect(Invitation.fromJson(goodJson)).to.be.an('object');
   });
 });
